server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Application server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.121.0",
    "psycopg[binary]>=3.3.0",
    "pyjwt>=2.10.1",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "FAST", "UP", "B", "SIM"]
